tp hopital

+++ JAVA/JEE/Exercices/Tp_hopital/src/main/java/com/example/tp_hopital/interfaces/Repository.java
@@ -10,7 +10,7 @@ public interface Repository<T> {
 
     boolean delete(T o);
 
-    T findById(int id);
+    T findById(Long id);
 
     List<T> findAll();
 }
